backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.3.0
python-multipart==0.0.20
transformers
torch
Pillow
timm

# PostgreSQL support
psycopg2-binary==2.9.9
pgvector==0.2.4

# Database migrations
alembic==1.13.1
